move ConfigId to its own class, fix building, other misc
