+++ src/main/java/dev/projectg/geyserhub/module/scoreboard/ScoreboardManager.java
@@ -1,6 +1,7 @@
 package dev.projectg.geyserhub.module.scoreboard;
 
 import dev.projectg.geyserhub.GeyserHubMain;
+import dev.projectg.geyserhub.config.ConfigId;
 import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -13,19 +14,18 @@ import java.util.Objects;
 
 
 public class ScoreboardManager {
-    public ScoreboardManager() {
-    }
+
+    public static final int refreshRate = Objects.requireNonNull(GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN)).getInt("Scoreboard.Refresh-rate");
 
     public static void addScoreboard() {
 
         for (Player all : Bukkit.getOnlinePlayers()) {
             createScoreboard(all);
         }
-
     }
 
     public static void createScoreboard(Player player) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         Scoreboard board = Objects.requireNonNull(Bukkit.getServer().getScoreboardManager()).getNewScoreboard();
         Objective objective = board.registerNewObjective("GeyserHub", "dummy", PlaceholderAPI.setPlaceholders(player, config.getString("Scoreboard.Title", "GeyserHub")));
