+++ src/main/java/dev/projectg/geyserhub/module/world/WorldSettings.java
@@ -1,6 +1,7 @@
 package dev.projectg.geyserhub.module.world;
 
 import dev.projectg.geyserhub.GeyserHubMain;
+import dev.projectg.geyserhub.config.ConfigId;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.configuration.file.FileConfiguration;
@@ -23,7 +24,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler
     public void onEntityDamage(EntityDamageEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!(event.getEntity() instanceof Player)) return;
 
@@ -44,7 +45,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onFoodChange(FoodLevelChangeEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-hunger-loss"))
             return;
@@ -55,7 +56,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler
     public void onFireSpread(BlockIgniteEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-block-fire-spread"))
             return;
@@ -65,7 +66,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler
     public void onBlockBurn(BlockBurnEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-disable-block-burn"))
             return;
@@ -74,7 +75,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onLeafDecay(LeavesDecayEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable_block-leaf-decay"))
             return;
@@ -83,7 +84,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onCreatureSpawn(CreatureSpawnEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-mob-spawning"))
             return;
@@ -93,7 +94,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onWeatherChange(WeatherChangeEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-weather-change"))
             return;
@@ -103,7 +104,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onEntityDamage(EntityDamageByEntityEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-player-pvp"))
             return;
@@ -113,7 +114,7 @@ public class WorldSettings implements Listener {
     }
     @EventHandler(priority = EventPriority.HIGH)
     public void onBlockBreak(BlockBreakEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-block-break")
                 || event.isCancelled())
@@ -129,7 +130,7 @@ public class WorldSettings implements Listener {
 
     @EventHandler(priority = EventPriority.HIGH)
     public void onBlockPlace(BlockPlaceEvent event) {
-        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration("config");
+        FileConfiguration config = GeyserHubMain.getInstance().getConfigManager().getFileConfiguration(ConfigId.MAIN);
         Objects.requireNonNull(config);
         if (!config.getBoolean("World-settings.disable-block-place")
                 || event.isCancelled())
